[#22]; refactor: 멤버 관리 페이지 코드 통합
